Second demo

+++ backend/JobList/Controllers/CompaniesController.cs
@@ -1,26 +1,35 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using JobList.Authorization;
 using JobList.BusinessLogic.Interfaces;
 using JobList.Common.DTOS;
+using JobList.Common.Errors;
+using JobList.Common.Pagination;
 using JobList.Common.Requests;
+using JobList.Common.Sorting;
+using Microsoft.AspNetCore.Authorization;
+using JobList.Common.UrlQuery;
 using Microsoft.AspNetCore.Mvc;
-
+using Newtonsoft.Json;
 
 namespace JobList.Controllers
 {
-    [Route("api/[controller]")]
+    [Route("[controller]")]
     [ApiController]
     public class CompaniesController : Controller
     {
+        private readonly IAuthorizationService _authorizationService;
         private ICompaniesService _companiesService;
 
-        public CompaniesController(ICompaniesService companiesService)
+        public CompaniesController(ICompaniesService companiesService, IAuthorizationService authorizationService)
         {
+            _authorizationService = authorizationService;
             _companiesService = companiesService;
         }
 
         // GET: /companies
+        [AllowAnonymous]
         [HttpGet]
         public virtual async Task<ActionResult<IEnumerable<CompanyDTO>>> Get()
         {
@@ -33,7 +42,30 @@ namespace JobList.Controllers
             return Ok(dtos);
         }
 
+        [AllowAnonymous]
+        [HttpGet("filtered")]
+        public virtual async Task<ActionResult<IEnumerable<RecruiterDTO>>> Get([FromQuery]SearchingUrlQuery searchingUrlQuery = null, [FromQuery]SortingUrlQuery sortingUrlQuery = null,
+                                                                            [FromQuery]PaginationUrlQuery paginationUrlQuery = null)
+        {
+            var dtos = await _companiesService.GetFilteredEntitiesAsync(searchingUrlQuery, sortingUrlQuery, paginationUrlQuery);
+            if (!dtos.Any())
+            {
+                return NoContent();
+            }
 
+            var pageInfo = new PageInfo()
+            {
+                PageNumber = paginationUrlQuery.PageNumber,
+                PageSize = paginationUrlQuery.PageSize,
+                TotalRecords = _companiesService.TotalRecords
+            };
+
+            Response.Headers.Add("X-Pagination", JsonConvert.SerializeObject(pageInfo));
+
+            return Ok(dtos);
+        }
+
+        [AllowAnonymous]
         [HttpGet("{id}")]
         public virtual async Task<ActionResult<CompanyDTO>> GetById(int id)
         {
@@ -47,27 +79,45 @@ namespace JobList.Controllers
         }
 
         // POST: /companies
-        [HttpPost]
-        public virtual async Task<ActionResult<CompanyDTO>> Create([FromBody] CompanyRequest request)
+        [AllowAnonymous]
+        [HttpPost("register")]
+        public virtual async Task<ActionResult<CompanyDTO>> Register([FromBody] CompanyRequest request)
         {
             if (!ModelState.IsValid)
             {
                 return BadRequest(ModelState);
             }
-
-            var dtos = await _companiesService.CreateEntityAsync(request);
-            if (dtos == null)
+            try
             {
-                return StatusCode(500);
-            }
+                var dtos = await _companiesService.CreateEntityAsync(request);
 
-            return CreatedAtAction("GetById", new { id = dtos.Id }, dtos);
+                if (dtos == null)
+                {
+                    return StatusCode(500, "Sth went wrong. Please try again!");
+                }
+
+                return CreatedAtAction("GetById", new { id = dtos.Id }, dtos);
+            }
+            catch (HttpStatusCodeException ex)
+            {
+                return BadRequest(ex.Message);
+            }
         }
 
         // PUT: /companies/:id
+        [Authorize(Roles = "company, admin")]
         [HttpPut("{id}")]
-        public virtual async Task<ActionResult> Update([FromRoute]int id, [FromBody]CompanyRequest request)
+        public virtual async Task<ActionResult> Update([FromRoute]int id, [FromBody]CompanyUpdateRequest request)
         {
+            var isAuthorized = await _authorizationService
+                    .AuthorizeAsync(User, id, UserOperations.Update);
+
+            if (!isAuthorized.Succeeded)
+            {
+                return Forbid();
+            }
+
+
             if (!ModelState.IsValid)
             {
                 return BadRequest(ModelState);
@@ -83,9 +133,18 @@ namespace JobList.Controllers
         }
 
         // DELETE: /companies/:id
+        [Authorize(Roles = "company, admin")]
         [HttpDelete("{id}")]
         public virtual async Task<ActionResult> Delete(int id)
         {
+            var isAuthorized = await _authorizationService
+                    .AuthorizeAsync(User, id, UserOperations.Delete);
+
+            if (!isAuthorized.Succeeded)
+            {
+                return Forbid();
+            }
+
             var result = await _companiesService.DeleteEntityByIdAsync(id);
             if (!result)
             {
