+++ backend/JobList/Controllers/EducationPeriodsController.cs
@@ -1,26 +1,30 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using JobList.Authorization;
 using JobList.BusinessLogic.Interfaces;
 using JobList.Common.DTOS;
 using JobList.Common.Requests;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
-
 
 namespace JobList.Controllers
 {
-    [Route("api/[controller]")]
+    [Route("[controller]")]
     [ApiController]
     public class EducationPeriodsController : Controller
     {
+        private readonly IAuthorizationService _authorizationService;
         private IEducationPeriodsService _educationPeriodsService;
 
-        public EducationPeriodsController(IEducationPeriodsService educationPeriodsService)
+        public EducationPeriodsController(IEducationPeriodsService educationPeriodsService, IAuthorizationService authorizationService)
         {
+            _authorizationService = authorizationService;
             _educationPeriodsService = educationPeriodsService;
         }
 
         // GET: /educationPeriods
+        [AllowAnonymous]
         [HttpGet]
         public virtual async Task<ActionResult<IEnumerable<EducationPeriodDTO>>> Get()
         {
@@ -33,7 +37,7 @@ namespace JobList.Controllers
             return Ok(dtos);
         }
 
-
+        [AllowAnonymous]
         [HttpGet("{id}")]
         public virtual async Task<ActionResult<EducationPeriodDTO>> GetById(int id)
         {
@@ -47,6 +51,7 @@ namespace JobList.Controllers
         }
 
         // POST: /educationPeriods
+        [Authorize(Roles = "employee, admin")]
         [HttpPost]
         public virtual async Task<ActionResult<EducationPeriodDTO>> Create([FromBody] EducationPeriodRequest request)
         {
@@ -65,9 +70,18 @@ namespace JobList.Controllers
         }
 
         // PUT: /educationPeriods/:id
+        [Authorize(Roles = "employee, admin")]
         [HttpPut("{id}")]
         public virtual async Task<ActionResult> Update([FromRoute]int id, [FromBody]EducationPeriodRequest request)
         {
+            var isAuthorized = await _authorizationService
+                    .AuthorizeAsync(User, request.ResumeId, UserOperations.Update);
+
+            if (!isAuthorized.Succeeded)
+            {
+                return Forbid();
+            }
+
             if (!ModelState.IsValid)
             {
                 return BadRequest(ModelState);
@@ -82,10 +96,34 @@ namespace JobList.Controllers
             return NoContent();
         }
 
+
+        [HttpGet("{id}")]
+        public virtual async Task<ActionResult<IEnumerable<EducationPeriodDTO>>> GetEducationPeriodsByResumeId(int id)
+        {
+            var dtos = await _educationPeriodsService.GetEducationPeriodsByResumeId(id);
+            if (!dtos.Any())
+            {
+                return NoContent();
+            }
+            return Ok(dtos);
+        }
+
+
         // DELETE: /educationPeriods/:id
+        [Authorize(Roles = "employee, admin")]
         [HttpDelete("{id}")]
         public virtual async Task<ActionResult> Delete(int id)
         {
+            var entity = await _educationPeriodsService.GetEntityByIdAsync(id);
+
+            var isAuthorized = await _authorizationService
+                    .AuthorizeAsync(User, entity.ResumeId, UserOperations.Delete);
+
+            if (!isAuthorized.Succeeded)
+            {
+                return Forbid();
+            }
+
             var result = await _educationPeriodsService.DeleteEntityByIdAsync(id);
             if (!result)
             {
