+++ backend/JobList/Controllers/FavoriteVacanciesController.cs
@@ -1,26 +1,31 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using JobList.Authorization;
 using JobList.BusinessLogic.Interfaces;
 using JobList.Common.DTOS;
 using JobList.Common.Requests;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
 
 namespace JobList.Controllers
 {
-    [Route("api/[controller]")]
+    [Route("[controller]")]
     [ApiController]
     public class FavoriteVacanciesController : Controller
     {
+        private readonly IAuthorizationService _authorizationService;
         private IFavoriteVacanciesService _favoriteVacanciesService;
 
-        public FavoriteVacanciesController(IFavoriteVacanciesService favoriteVacanciesService)
+        public FavoriteVacanciesController(IFavoriteVacanciesService favoriteVacanciesService, IAuthorizationService authorizationService)
         {
+            _authorizationService = authorizationService;
             _favoriteVacanciesService = favoriteVacanciesService;
         }
 
         // GET: /favoriteVacancies
+        [AllowAnonymous]
         [HttpGet]
         public virtual async Task<ActionResult<IEnumerable<FavoriteVacancyDTO>>> Get()
         {
@@ -33,7 +38,7 @@ namespace JobList.Controllers
             return Ok(dtos);
         }
 
-
+        [AllowAnonymous]
         [HttpGet("{id}")]
         public virtual async Task<ActionResult<FavoriteVacancyDTO>> GetById(int id)
         {
@@ -47,6 +52,7 @@ namespace JobList.Controllers
         }
 
         // POST: /favoriteVacancies
+        [Authorize(Roles = "employee, admin")]
         [HttpPost]
         public virtual async Task<ActionResult<FavoriteVacancyDTO>> Create([FromBody] FavoriteVacancyRequest request)
         {
@@ -65,9 +71,18 @@ namespace JobList.Controllers
         }
 
         // PUT: /favoriteVacancies/:id
+        [Authorize(Roles = "employee, admin")]
         [HttpPut("{id}")]
         public virtual async Task<ActionResult> Update([FromRoute]int id, [FromBody]FavoriteVacancyRequest request)
         {
+            var isAuthorized = await _authorizationService
+                    .AuthorizeAsync(User, request.EmployeeId, UserOperations.Update);
+
+            if (!isAuthorized.Succeeded)
+            {
+                return Forbid();
+            }
+
             if (!ModelState.IsValid)
             {
                 return BadRequest(ModelState);
@@ -83,9 +98,20 @@ namespace JobList.Controllers
         }
 
         // DELETE: /favoriteVacancies/:id
+        [Authorize(Roles = "employee, admin")]
         [HttpDelete("{id}")]
         public virtual async Task<ActionResult> Delete(int id)
         {
+            var entity = await _favoriteVacanciesService.GetEntityByIdAsync(id);
+
+            var isAuthorized = await _authorizationService
+                    .AuthorizeAsync(User, entity.EmployeerId, UserOperations.Delete);
+
+            if (!isAuthorized.Succeeded)
+            {
+                return Forbid();
+            }
+
             var result = await _favoriteVacanciesService.DeleteEntityByIdAsync(id);
             if (!result)
             {
